feature/home: refactor folder remove components folder in components folder
