+++ src/app/modules/home/components/short-link/result-short/result-short.component.html
