+++ src/app/modules/home/components/statistic-list/statistic-list.component.ts
