+++ src/app/modules/home/home-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomePageComponent } from './components/components/home-page.component';
+import { HomePageComponent } from './components/home-page.component';
 
 const routes: Routes = [
   {path: '', component: HomePageComponent}
