+++ src/app/modules/home/home.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { HomeRoutingModule } from './home-routing.module';
-import { HomePageComponent } from './components/components/home-page.component';
+import { HomePageComponent } from './components/home-page.component';
 import { CardStepComponent } from 'projects/shortly-lib/src/lib/components/card-step/card-step.component';
-import { StatisticListComponent } from './components/components/statistic-list/statistic-list.component';
-import { ShortLinkComponent } from './components/components/short-link/short-link.component';
-import { FormShortComponent } from './components/components/short-link/form-short/form-short.component';
-import { ResultShortComponent } from './components/components/short-link/result-short/result-short.component';
+import { StatisticListComponent } from './components/statistic-list/statistic-list.component';
+import { ShortLinkComponent } from './components/short-link/short-link.component';
+import { FormShortComponent } from './components/short-link/form-short/form-short.component';
+import { ResultShortComponent } from './components/short-link/result-short/result-short.component';
 
 
 @NgModule({
